backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
google-generativeai==0.8.3
requests==2.31.0
Pillow==10.1.0
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
